+++ FakeServer/Startup.cs
@@ -12,6 +12,7 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.PlatformAbstractions;
 using Swashbuckle.AspNetCore.Swagger;
 using Swashbuckle.AspNetCore.SwaggerUI;
@@ -30,6 +31,19 @@ namespace FakeServer
 
         public void ConfigureServices(IServiceCollection services)
         {
+            var folder = Configuration["staticFolder"];
+
+            if (!string.IsNullOrEmpty(folder))
+            {
+                services.AddSpaStaticFiles((spa) =>
+                {
+                    spa.RootPath = folder;
+                });
+
+                // No need to define anything else as this can only be used as a SPA server
+                return;
+            }
+
             var jsonFilePath = Path.Combine(Configuration["currentPath"], Configuration["file"]);
             services.AddSingleton<IDataStore>(new DataStore(jsonFilePath, reloadBeforeGetCollection: Configuration.GetValue<bool>("Common:EagerDataReload")));
             services.AddSingleton<IMessageBus, MessageBus>();
@@ -92,6 +106,28 @@ namespace FakeServer
 
         public void Configure(IApplicationBuilder app, IHostingEnvironment env, IApplicationLifetime appLifetime)
         {
+            var folder = Configuration["staticFolder"];
+
+            app.UseDefaultFiles();
+
+            if (string.IsNullOrEmpty(folder))
+            {
+                app.UseStaticFiles();
+            }
+            else
+            {
+                app.UseSpa(spa =>
+                {
+                    spa.ApplicationBuilder.UseSpaStaticFiles(new StaticFileOptions
+                    {
+                        FileProvider = new PhysicalFileProvider(folder)
+                    });
+                });
+
+                // No need to define anything else as this can only be used as a SPA server
+                return;
+            }
+
             app.UseCors("AllowAnyPolicy");
 
             app.UseMiddleware<HttpOptionsMiddleware>();
@@ -121,9 +157,6 @@ namespace FakeServer
             {
                 TokenConfiguration.UseTokenProviderMiddleware(app);
             }
-
-            app.UseDefaultFiles();
-            app.UseStaticFiles();
 
             if (Configuration.GetValue<bool>("Caching:ETag:Enabled"))
             {