+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
 ### [Unreleased]
-* Add support for paging with page and per_page
-
+* ADDED: Support for paging with page and per_page
+* ADDED: User defined location for the static files (including SPAs)
+ 
 ### [0.8.0] - 2018-10-14
 * Release as a dotnet tool
 * Use executable location as a base path for settings files